# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted webmail application API backed by IMAP and SMTP.",
        "title": "Go-Cubemail API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/folders": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the complete hierarchical list of IMAP mailboxes/folders with unseen/total counts and tree structure metadata.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailbox"
                ],
                "summary": "List IMAP folders",
                "responses": {
                    "200": {
                        "description": "Success response containing the list of folders",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/imap.MailboxInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a new IMAP mailbox folder, optionally nested under a parent folder.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailbox"
                ],
                "summary": "Create subfolder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Parent folder name (for nesting)",
                        "name": "parent",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Hierarchy delimiter (default: /)",
                        "name": "delim",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status and full folder name",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "name is required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "IMAP or connection error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/folders/delete": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Recursively deletes an IMAP folder and all its subfolders.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailbox"
                ],
                "summary": "Delete folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder name to delete",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "name missing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "IMAP error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/folders/rename": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Renames an existing IMAP mailbox folder.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailbox"
                ],
                "summary": "Rename folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Current folder name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New folder name",
                        "name": "newname",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "name or newname missing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "IMAP error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/folders/{name}/count": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the number of unseen (unread) messages in the specified folder.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailbox"
                ],
                "summary": "Unread message count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "unseen count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    }
                }
            }
        },
...
        "/mail/{mailbox}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns a paginated list of email message envelopes for the specified folder, ordered newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mailbox"
                ],
                "summary": "List emails in mailbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailbox folder name (e.g. INBOX, Drafts, Sent, Trash)",
                        "name": "mailbox",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (defaults to 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response containing mailbox, messages, page, and total count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Mailbox not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Permanently expunges all messages in Trash. If the given mailbox is not Trash, moves all its messages to Trash instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mail"
                ],
                "summary": "Empty trash / clear folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailbox folder name",
                        "name": "mailbox",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/mail/{mailbox}/{uid}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Fetches the full message: envelope, sanitized HTML body, plain-text body, and attachment list. Marks the message as seen.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mail"
                ],
                "summary": "Read message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailbox folder name",
                        "name": "mailbox",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Message UID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message envelope, html_body, plain_body, attachments, calendar info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "invalid uid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "message not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Moves a message to Trash. If already in Trash, permanently expunges it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mail"
                ],
                "summary": "Delete message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailbox folder name",
                        "name": "mailbox",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Message UID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid uid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "imap.MailboxInfo": {
            "type": "object",
            "properties": {
                "delim": {
                    "type": "string"
                },
                "depth": {
                    "description": "nesting level (0 = top-level)",
                    "type": "integer"
                },
                "displayName": {
                    "description": "last path segment (used for tree display)",
                    "type": "string"
                },
                "hasChildren": {
                    "description": "true if at least one subfolder exists under this folder",
                    "type": "boolean"
                },
                "iconType": {
                    "description": "\"inbox\" | \"drafts\" | \"sent\" | \"trash\" | \"junk\" | \"folder\"",
                    "type": "string"
                },
                "isSystem": {
                    "description": "true for INBOX, Sent, Drafts, Trash — cannot be renamed or deleted",
                    "type": "boolean"
                },
                "isTrash": {
                    "type": "boolean"
                },
                "messages": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "noSelect": {
                    "description": "true when the folder is a hierarchy placeholder (\\Noselect)",
                    "type": "boolean"
                },
                "paddingLeft": {
                    "description": "left-padding in px for tree indentation in the UI",
                    "type": "integer"
                },
                "parentName": {
                    "description": "full name of the parent folder; empty for top-level",
                    "type": "string"
                },
                "unseen": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "The session cookie 'gorc_session' is required for authenticated endpoints.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}